Moves test to light
